requirements.txt
streamlit
faiss-cpu
PyMuPDF
spacy
sentence-transformers
requests

